+++ lib/flutter_brand_palettes.dart
@@ -3,3 +3,4 @@ export 'src/google.dart';
 export 'src/instagram.dart';
 export 'src/palette.dart';
 export 'src/palette_rgb.dart';
+export 'src/whatsapp.dart';